refactor: remove comment service
