+++ src/appv2/features/comment/controllers.rs
@@ -1,12 +1,9 @@
-use super::{
-    presenters::{MultipleCommentsResponse, SingleCommentResponse},
-    request, service,
-};
+use super::request;
 use crate::appv2::drivers::middlewares::auth;
 use crate::appv2::drivers::middlewares::state::AppState;
 use crate::utils::api::ApiResponse;
 use crate::utils::uuid;
-use actix_web::{web, HttpRequest, HttpResponse};
+use actix_web::{web, HttpRequest};
 
 type ArticleIdSlug = String;
 type CommentIdSlug = String;
@@ -17,10 +14,6 @@ pub async fn index(state: web::Data<AppState>, req: HttpRequest) -> ApiResponse 
         .di_container
         .comment_usecase
         .fetch_comments_list(&current_user)
-    // let conn = &mut state.get_conn()?;
-    // let list = service::fetch_comments_list(conn, &current_user)?;
-    // let res = MultipleCommentsResponse::from(list);
-    // Ok(HttpResponse::Ok().json(res))
 }
 
 pub async fn create(
@@ -29,19 +22,13 @@ pub async fn create(
     path: web::Path<ArticleIdSlug>,
     form: web::Json<request::CreateCommentRequest>,
 ) -> ApiResponse {
-    let conn = &mut state.get_conn()?;
     let current_user = auth::get_current_user(&req)?;
     let article_title_slug = path.into_inner();
-    let (comment, profile) = service::create(
-        conn,
-        &service::CreateCommentService {
-            body: form.comment.body.to_owned(),
-            article_title_slug,
-            author: current_user,
-        },
-    )?;
-    let res = SingleCommentResponse::from((comment, profile));
-    Ok(HttpResponse::Ok().json(res))
+    let body = form.comment.body.to_owned();
+    state
+        .di_container
+        .comment_usecase
+        .create(body, article_title_slug, current_user)
 }
 
 pub async fn delete(
