+++ src/appv2/features/comment/mod.rs
@@ -3,5 +3,4 @@ pub mod entities;
 pub mod presenters;
 pub mod repositories;
 pub mod request;
-pub mod service;
 pub mod usecases;
